Refs: #11315: renamed tests and test the insert of a participant linked to no process
