+++ test/unittest/Database/DatabaseLoadTests.cpp
@@ -12,1014 +12,1092 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <fstream>
+
 #include "gtest/gtest.h"
+
+#include <fastdds-statistics-backend/exception/Exception.hpp>
+#include <fastdds-statistics-backend/types/JSONTags.h>
+#include <fastdds-statistics-backend/types/types.hpp>
 
 #include <database/database.hpp>
 
 using namespace eprosima::statistics_backend;
 using namespace eprosima::statistics_backend::database;
 
+constexpr const char* DESCRIPTION_TAG = "description";
+
+constexpr const char* EMPTY_DUMP_FILE = "resources/empty_dump.json";
+constexpr const char* EMPTY_ENTITIES_DUMP_FILE = "resources/empty_entities_dump.json";
+constexpr const char* SIMPLE_DUMP_FILE = "resources/simple_dump.json";
+constexpr const char* COMPLEX_DUMP_FILE = "resources/complex_dump.json";
+
+DatabaseDump load_file(
+        std::string filename)
+{
+    // Check if the file exists
+    std::ifstream file(filename);
+    if (!file.good())
+    {
+        throw BadParameter("File " + filename + " does not exist");
+    }
+
+    // Get the json file
+    DatabaseDump dump;
+    file >> dump;
+
+    // Erase the description tag if existing
+    if (dump.contains(DESCRIPTION_TAG))
+    {
+        dump.erase(DESCRIPTION_TAG);
+    }
+
+    return dump;
+}
+
+// Test the load of a dump database without any entity
+TEST(database_load_tests, load_and_dump_empty_database)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(EMPTY_DUMP_FILE);
+
+    // Create database
+    Database db;
+
+    // Load jump in database
+    db.load_database(dump);
+
+    // Dump loaded database
+    DatabaseDump loadedDump = db.dump_database();
+
+    // Compare two dumps
+    ASSERT_EQ(dump, loadedDump);
+}
+
+// Test the load of a dump database with one entity of each kind
+TEST(database_load_tests, load_and_dump_empty_entities_database)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(EMPTY_ENTITIES_DUMP_FILE);
+
+    // Create database
+    Database db;
+
+    // Load jump in database
+    db.load_database(dump);
+
+    // Dump loaded database
+    DatabaseDump loadedDump = db.dump_database();
+
+    // Compare two dumps
+    ASSERT_EQ(dump[HOST_CONTAINER_TAG], loadedDump[HOST_CONTAINER_TAG]);
+    ASSERT_EQ(dump[USER_CONTAINER_TAG], loadedDump[USER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PROCESS_CONTAINER_TAG], loadedDump[PROCESS_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DOMAIN_CONTAINER_TAG], loadedDump[DOMAIN_CONTAINER_TAG]);
+    ASSERT_EQ(dump[TOPIC_CONTAINER_TAG], loadedDump[TOPIC_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PARTICIPANT_CONTAINER_TAG], loadedDump[PARTICIPANT_CONTAINER_TAG]);
+    ASSERT_EQ(dump[LOCATOR_CONTAINER_TAG], loadedDump[LOCATOR_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAWRITER_CONTAINER_TAG], loadedDump[DATAWRITER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAREADER_CONTAINER_TAG], loadedDump[DATAREADER_CONTAINER_TAG]);
+
+    ASSERT_EQ(dump, loadedDump);
+}
+
+// Test the load of a dump database with one entity of each kind and one data of each kind
+TEST(database_load_tests, load_and_dump_simple_database)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(SIMPLE_DUMP_FILE);
+
+    // Create database
+    Database db;
+
+    // Load jump in database
+    db.load_database(dump);
+
+    // Dump loaded database
+    DatabaseDump loadedDump = db.dump_database();
+
+    // Compare two dumps
+    ASSERT_EQ(dump[HOST_CONTAINER_TAG], loadedDump[HOST_CONTAINER_TAG]);
+    ASSERT_EQ(dump[USER_CONTAINER_TAG], loadedDump[USER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PROCESS_CONTAINER_TAG], loadedDump[PROCESS_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DOMAIN_CONTAINER_TAG], loadedDump[DOMAIN_CONTAINER_TAG]);
+    ASSERT_EQ(dump[TOPIC_CONTAINER_TAG], loadedDump[TOPIC_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PARTICIPANT_CONTAINER_TAG], loadedDump[PARTICIPANT_CONTAINER_TAG]);
+    ASSERT_EQ(dump[LOCATOR_CONTAINER_TAG], loadedDump[LOCATOR_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAWRITER_CONTAINER_TAG], loadedDump[DATAWRITER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAREADER_CONTAINER_TAG], loadedDump[DATAREADER_CONTAINER_TAG]);
+
+    ASSERT_EQ(dump, loadedDump);
+}
+
+// Test the load of a dump database with three entities of each kind and three datas of each kind
+TEST(database_load_tests, load_and_dump_complex_database)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(COMPLEX_DUMP_FILE);
+
+    // Create database
+    Database db;
+
+    // Load jump in database
+    db.load_database(dump);
+
+    // Dump loaded database
+    DatabaseDump loadedDump = db.dump_database();
+
+    // Compare two dumps
+    ASSERT_EQ(dump[HOST_CONTAINER_TAG], loadedDump[HOST_CONTAINER_TAG]);
+    ASSERT_EQ(dump[USER_CONTAINER_TAG], loadedDump[USER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PROCESS_CONTAINER_TAG], loadedDump[PROCESS_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DOMAIN_CONTAINER_TAG], loadedDump[DOMAIN_CONTAINER_TAG]);
+    ASSERT_EQ(dump[TOPIC_CONTAINER_TAG], loadedDump[TOPIC_CONTAINER_TAG]);
+    ASSERT_EQ(dump[PARTICIPANT_CONTAINER_TAG], loadedDump[PARTICIPANT_CONTAINER_TAG]);
+    ASSERT_EQ(dump[LOCATOR_CONTAINER_TAG], loadedDump[LOCATOR_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAWRITER_CONTAINER_TAG], loadedDump[DATAWRITER_CONTAINER_TAG]);
+    ASSERT_EQ(dump[DATAREADER_CONTAINER_TAG], loadedDump[DATAREADER_CONTAINER_TAG]);
+
+    ASSERT_EQ(dump, loadedDump);
+}
+
+
+// Check that the load of the 'dump 'without 'key' in 'container', throws an exception
+void check_no_key(
+        DatabaseDump& dump,
+        DatabaseDump& container,
+        std::string const& key,
+        bool error = true)
+{
+    Database db;
+    container.erase(key);
+    if (error)
+    {
+        ASSERT_ANY_THROW(db.load_database(dump));
+    }
+    else
+    {
+        ASSERT_NO_THROW(db.load_database(dump));
+    }
+}
+
+void check_entity_no_generic_keys(
+        DatabaseDump& dump,
+        std::string const& key)
+{
+    // No entities key
+    {
+        DatabaseDump dumpCopy = dump;
+        check_no_key(dumpCopy, dumpCopy, key);
+    }
+
+    // No entity key
+    {
+        DatabaseDump dumpCopy = dump;
+        check_no_key(dumpCopy, dumpCopy[key], dumpCopy[key].begin().key());
+    }
+
+    // No entity name key
+    {
+        DatabaseDump dumpCopy = dump;
+        check_no_key(dumpCopy, dumpCopy[key][dumpCopy[key].begin().key()], NAME_INFO_TAG);
+    }
+}
+
+void check_entity_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy, dumpCopy[entityTag][dumpCopy[entityTag].begin().key()], keyTag);
+}
+
+void check_data_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy, dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG], keyTag);
+}
+
+void check_data_value_no_id_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy, dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag],
+            dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag].begin().key(), false);
+}
+
+void check_data_value_index_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag,
+        std::string const& data_param)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy,
+            dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag]
+                    .begin().value().begin().value(),
+            data_param);
+}
+
+void check_data_value_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag,
+        std::string const& data_param)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy,
+            dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag]
+                    .begin().value(),
+            data_param);
+}
+
+void check_data_value_index_last_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag,
+        std::string const& data_param)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy,
+            dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag]
+                    .begin().value(),
+            data_param);
+}
+
+void check_data_value_last_no_key(
+        DatabaseDump& dump,
+        std::string const& entityTag,
+        std::string const& keyTag,
+        std::string const& data_param)
+{
+    DatabaseDump dumpCopy = dump;
+    check_no_key(dumpCopy,
+            dumpCopy[entityTag][dumpCopy[entityTag].begin().key()][DATA_VALUE_DATA_TAG][keyTag],
+            data_param);
+}
+
+// Test the load of a corrupted database with erased keys
+TEST(database_load_tests, load_erased_keys)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(SIMPLE_DUMP_FILE);
+
+    constexpr const char* char_message = "test";
+    std::string string_message = "test";
+    ASSERT_ANY_THROW(throw CorruptedFile(char_message));
+    ASSERT_ANY_THROW(throw CorruptedFile(string_message));
+
+    // ------------ HOSTS ----------------
+
+    check_entity_no_generic_keys(dump, HOST_CONTAINER_TAG);
+
+    check_entity_no_key(dump, HOST_CONTAINER_TAG, USER_CONTAINER_TAG);
+
+    // ------------ USERS ----------------
+
+    check_entity_no_generic_keys(dump, USER_CONTAINER_TAG);
+
+    check_entity_no_key(dump, USER_CONTAINER_TAG, HOST_ENTITY_TAG);
+    check_entity_no_key(dump, USER_CONTAINER_TAG, PROCESS_CONTAINER_TAG);
+
+    // ------------ PROCESSES ----------------
+
+    check_entity_no_generic_keys(dump, PROCESS_CONTAINER_TAG);
+
+    check_entity_no_key(dump, PROCESS_CONTAINER_TAG, PID_INFO_TAG);
+    check_entity_no_key(dump, PROCESS_CONTAINER_TAG, USER_ENTITY_TAG);
+    check_entity_no_key(dump, PROCESS_CONTAINER_TAG, PARTICIPANT_CONTAINER_TAG);
+
+    // ------------ DOMAINS ----------------
+
+    check_entity_no_generic_keys(dump, DOMAIN_CONTAINER_TAG);
+
+    check_entity_no_key(dump, DOMAIN_CONTAINER_TAG, PARTICIPANT_CONTAINER_TAG);
+    check_entity_no_key(dump, DOMAIN_CONTAINER_TAG, TOPIC_CONTAINER_TAG);
+
+    // ------------ TOPICS ----------------
+
+    check_entity_no_generic_keys(dump, TOPIC_CONTAINER_TAG);
+
+    check_entity_no_key(dump, TOPIC_CONTAINER_TAG, DATA_TYPE_INFO_TAG);
+    check_entity_no_key(dump, TOPIC_CONTAINER_TAG, DOMAIN_ENTITY_TAG);
+    check_entity_no_key(dump, TOPIC_CONTAINER_TAG, DATAWRITER_CONTAINER_TAG);
+    check_entity_no_key(dump, TOPIC_CONTAINER_TAG, DATAREADER_CONTAINER_TAG);
+
+    // ------------ PARTICIPANTS ----------------
+
+    check_entity_no_generic_keys(dump, PARTICIPANT_CONTAINER_TAG);
+
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, GUID_INFO_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, QOS_INFO_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, PROCESS_ENTITY_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, DOMAIN_ENTITY_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATAWRITER_CONTAINER_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATAREADER_CONTAINER_TAG);
+    check_entity_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_VALUE_DATA_TAG);
+
+    // ------------ LOCATORS ----------------
+
+    check_entity_no_generic_keys(dump, LOCATOR_CONTAINER_TAG);
+
+    check_entity_no_key(dump, LOCATOR_CONTAINER_TAG, DATAWRITER_CONTAINER_TAG);
+    check_entity_no_key(dump, LOCATOR_CONTAINER_TAG, DATAREADER_CONTAINER_TAG);
+    check_entity_no_key(dump, LOCATOR_CONTAINER_TAG, DATA_VALUE_DATA_TAG);
+
+    // ------------ DATAWRITERS ----------------
+
+    check_entity_no_generic_keys(dump, DATAWRITER_CONTAINER_TAG);
+
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, GUID_INFO_TAG);
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, QOS_INFO_TAG);
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, PARTICIPANT_ENTITY_TAG);
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, TOPIC_ENTITY_TAG);
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, LOCATOR_CONTAINER_TAG);
+    check_entity_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_VALUE_DATA_TAG);
+
+    // ------------ DATAREADERS ----------------
+
+    check_entity_no_generic_keys(dump, DATAREADER_CONTAINER_TAG);
+
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, GUID_INFO_TAG);
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, QOS_INFO_TAG);
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, PARTICIPANT_ENTITY_TAG);
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, TOPIC_ENTITY_TAG);
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, LOCATOR_CONTAINER_TAG);
+    check_entity_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_VALUE_DATA_TAG);
+
+    // ------------ PARTICIPANT DATA ----------------
+
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_LAST_REPORTED_TAG);
+    check_data_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_LAST_REPORTED_TAG);
+
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG);
+    check_data_value_no_id_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG);
+
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG, DATA_VALUE_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG,
+            DATA_VALUE_REMOTE_ENTITY_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_DISCOVERY_TIME_TAG,
+            DATA_VALUE_DISCOVERED_TAG);
+    check_data_value_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_TAG,
+            DATA_VALUE_MAGNITUDE_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_TAG,
+            DATA_VALUE_MAGNITUDE_TAG);
+    check_data_value_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_PDP_PACKETS_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_EDP_PACKETS_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG,
+            DATA_VALUE_MAGNITUDE_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_index_last_no_key(dump, PARTICIPANT_CONTAINER_TAG, DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG,
+            DATA_VALUE_MAGNITUDE_TAG);
+
+    // ------------ DATAWRITER DATA ----------------
+
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_PUBLICATION_THROUGHPUT_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_SAMPLE_DATAS_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_FASTDDS_LATENCY_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_LAST_REPORTED_TAG);
+
+    check_data_value_no_id_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_SAMPLE_DATAS_TAG);
+    check_data_value_no_id_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_FASTDDS_LATENCY_TAG);
+
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_PUBLICATION_THROUGHPUT_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_PUBLICATION_THROUGHPUT_TAG, DATA_VALUE_DATA_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_SAMPLE_DATAS_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_SAMPLE_DATAS_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_index_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_FASTDDS_LATENCY_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_FASTDDS_LATENCY_TAG, DATA_VALUE_DATA_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_DATA_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_GAP_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_HEARTBEAT_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAWRITER_CONTAINER_TAG, DATA_KIND_RESENT_DATA_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+
+    // ------------ DATAREADER DATA ----------------
+
+    check_data_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_SUBSCRIPTION_THROUGHPUT_TAG);
+    check_data_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_TAG);
+    check_data_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_TAG);
+    check_data_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_LAST_REPORTED_TAG);
+    check_data_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_LAST_REPORTED_TAG);
+
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_SUBSCRIPTION_THROUGHPUT_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_SUBSCRIPTION_THROUGHPUT_TAG, DATA_VALUE_DATA_TAG);
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_TAG, DATA_VALUE_COUNT_TAG);
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_TAG, DATA_VALUE_COUNT_TAG);
+
+    check_data_value_last_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_ACKNACK_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+
+    check_data_value_last_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_last_no_key(dump, DATAREADER_CONTAINER_TAG, DATA_KIND_NACKFRAG_COUNT_LAST_REPORTED_TAG,
+            DATA_VALUE_COUNT_TAG);
+
+    // ------------ LOCATOR DATA ----------------
+
+    check_data_no_key(dump, LOCATOR_CONTAINER_TAG, DATA_KIND_NETWORK_LATENCY_TAG);
+
+    check_data_value_no_id_key(dump, LOCATOR_CONTAINER_TAG, DATA_KIND_NETWORK_LATENCY_TAG);
+    check_data_value_index_no_key(dump, LOCATOR_CONTAINER_TAG, DATA_KIND_NETWORK_LATENCY_TAG, DATA_VALUE_SRC_TIME_TAG);
+    check_data_value_index_no_key(dump, LOCATOR_CONTAINER_TAG, DATA_KIND_NETWORK_LATENCY_TAG, DATA_VALUE_DATA_TAG);
+}
+
+void check_and_restore(
+        DatabaseDump& dump,
+        DatabaseDump& container,
+        std::vector<DatabaseDump>const& wrongValues,
+        std::vector<DatabaseDump>const& correctValues)
+{
+    // Wrong
+    for (auto wrongValue: wrongValues)
+    {
+        DatabaseDump aux = DatabaseDump(container);
+        container = wrongValue;
+        Database db;
+        ASSERT_ANY_THROW(db.load_database(dump));
+        container = aux;
+    }
+
+    for (auto correctValue: correctValues)
+    {
+        DatabaseDump aux = DatabaseDump(container);
+        container = correctValue;
+        Database db;
+        ASSERT_NO_THROW(db.load_database(dump));
+        container = aux;
+    }
+}
+
+void check_is_string(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array(
+                                                 {"array"}), true, 1234, 34.12};
+    std::vector<DatabaseDump> correctValues = {"qwerty", "1234", "34.12"};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_id(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             true, 1234, 34.12,
+                                             "qwerty", "1234", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_object(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {};
+    std::vector<DatabaseDump> correctValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1,
+                                                                                                                     2,
+                                                                                                                     3,
+                                                                                                                     4}),
+                                               true, 1234, 34.12,
+                                               "qwerty", "1234", "34.12"};
+
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_string_int(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             true, 1234, 34.12,
+                                             "qwerty", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {"1234"};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_bool(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             1234, 34.12,
+                                             "qwerty", "1234", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {true};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_uint(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             true,
+                                             "qwerty", "1234", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {1234, 34.12};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_int(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             "qwerty", "1234", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {true, 1234, 34.12};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+void check_is_double(
+        DatabaseDump& dump,
+        DatabaseDump& container)
+{
+    std::vector<DatabaseDump> wrongValues = {DatabaseDump::object({{"one", 1}, {"two", 2} }), DatabaseDump::array({1, 2,
+                                                                                                                   3,
+                                                                                                                   4}),
+                                             true,
+                                             "qwerty", "1234", "34.12"};
+
+    std::vector<DatabaseDump> correctValues = {1234, 34.12};
+    check_and_restore(dump, container, wrongValues, correctValues);
+}
+
+// Test the load of a database with wrong value types
+TEST(database_load_tests, load_wrong_values)
+{
+    // Read JSON
+    DatabaseDump dump = load_file(SIMPLE_DUMP_FILE);
+
+    // object, string, float, int, array, bool
+
+    // Host
+    check_is_string(dump, dump[HOST_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_id(dump, dump[HOST_CONTAINER_TAG].begin().value()[USER_CONTAINER_TAG]);
+
+    // User
+    check_is_string(dump, dump[USER_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_id(dump, dump[USER_CONTAINER_TAG].begin().value()[HOST_ENTITY_TAG]);
+    check_is_id(dump, dump[USER_CONTAINER_TAG].begin().value()[PROCESS_CONTAINER_TAG]);
+
+    // Process
+    check_is_string(dump, dump[PROCESS_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_string(dump, dump[PROCESS_CONTAINER_TAG].begin().value()[PID_INFO_TAG]);
+    check_is_id(dump, dump[PROCESS_CONTAINER_TAG].begin().value()[USER_ENTITY_TAG]);
+    check_is_id(dump, dump[PROCESS_CONTAINER_TAG].begin().value()[PARTICIPANT_CONTAINER_TAG]);
+
+    // Domain
+    check_is_string(dump, dump[DOMAIN_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_id(dump, dump[DOMAIN_CONTAINER_TAG].begin().value()[PARTICIPANT_CONTAINER_TAG]);
+    check_is_id(dump, dump[DOMAIN_CONTAINER_TAG].begin().value()[TOPIC_CONTAINER_TAG]);
+
+    // Topic
+    check_is_string(dump, dump[TOPIC_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_string(dump, dump[TOPIC_CONTAINER_TAG].begin().value()[DATA_TYPE_INFO_TAG]);
+    check_is_id(dump, dump[TOPIC_CONTAINER_TAG].begin().value()[DOMAIN_ENTITY_TAG]);
+    check_is_id(dump, dump[TOPIC_CONTAINER_TAG].begin().value()[DATAWRITER_CONTAINER_TAG]);
+    check_is_id(dump, dump[TOPIC_CONTAINER_TAG].begin().value()[DATAREADER_CONTAINER_TAG]);
+
+    // Participant
+    check_is_string(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_string(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[GUID_INFO_TAG]);
+    check_is_object(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[QOS_INFO_TAG]);
+    check_is_id(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[PROCESS_ENTITY_TAG]);
+    check_is_id(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DOMAIN_ENTITY_TAG]);
+    check_is_id(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATAWRITER_CONTAINER_TAG]);
+    check_is_id(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATAREADER_CONTAINER_TAG]);
+
+    // Locator
+    check_is_string(dump, dump[LOCATOR_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_id(dump, dump[LOCATOR_CONTAINER_TAG].begin().value()[DATAWRITER_CONTAINER_TAG]);
+    check_is_id(dump, dump[LOCATOR_CONTAINER_TAG].begin().value()[DATAREADER_CONTAINER_TAG]);
+
+    // Datawriter
+    check_is_string(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_string(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[GUID_INFO_TAG]);
+    check_is_object(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[QOS_INFO_TAG]);
+    check_is_id(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[PARTICIPANT_ENTITY_TAG]);
+    check_is_id(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[TOPIC_ENTITY_TAG]);
+    check_is_id(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[LOCATOR_CONTAINER_TAG]);
+
+    // Datareader
+    check_is_string(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[NAME_INFO_TAG]);
+    check_is_string(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[GUID_INFO_TAG]);
+    check_is_object(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[QOS_INFO_TAG]);
+    check_is_id(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[PARTICIPANT_ENTITY_TAG]);
+    check_is_id(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[TOPIC_ENTITY_TAG]);
+    check_is_id(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[LOCATOR_CONTAINER_TAG]);
+
+    /* _________________DATA_____________________ */
+
+    // --------------------- Participants ---------------------
+
+    // discovery_time
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DISCOVERY_TIME_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DISCOVERY_TIME_TAG].begin().value().begin().value()[DATA_VALUE_TIME_TAG]);
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DISCOVERY_TIME_TAG].begin().value().begin().value()[DATA_VALUE_REMOTE_ENTITY_TAG]);
+    check_is_bool(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DISCOVERY_TIME_TAG].begin().value().begin().value()[DATA_VALUE_DISCOVERED_TAG]);
+
+    // pdp_packets
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PDP_PACKETS_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PDP_PACKETS_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // edp_packets
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_EDP_PACKETS_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_EDP_PACKETS_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // rtps_packets_sent
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_SENT_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_SENT_TAG].begin().value().begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // rtps_bytes_sent
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_TAG].begin().value().begin().value()[DATA_VALUE_MAGNITUDE_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_TAG].begin().value().begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // rtps_packets_lost
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_LOST_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_LOST_TAG].begin().value().begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // rtps_bytes_lost
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_TAG].begin().value().begin().value()[DATA_VALUE_MAGNITUDE_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_TAG].begin().value().begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_edp_packets
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_EDP_PACKETS_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_EDP_PACKETS_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_pdp_packets
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PDP_PACKETS_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PDP_PACKETS_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_rtps_bytes_lost
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_MAGNITUDE_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_LOST_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_rtps_bytes_sent
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_MAGNITUDE_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_BYTES_SENT_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_rtps_packets_lost
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_LOST_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_rtps_packets_sent
+    check_is_string_int(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[PARTICIPANT_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RTPS_PACKETS_SENT_LAST_REPORTED_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+
+    // --------------------- Datawriters ---------------------
+
+    // publication_throughput
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PUBLICATION_THROUGHPUT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_double(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_PUBLICATION_THROUGHPUT_TAG].begin().value()[DATA_VALUE_DATA_TAG]);
+
+    // resent_datas
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RESENT_DATA_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RESENT_DATA_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // heartbeat_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_HEARTBEAT_COUNT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_HEARTBEAT_COUNT_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // gap_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_GAP_COUNT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_GAP_COUNT_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // data_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DATA_COUNT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DATA_COUNT_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // samples_datas
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_SAMPLE_DATAS_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_SAMPLE_DATAS_TAG].begin().value().begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // history2history_latency
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_FASTDDS_LATENCY_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_double(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_FASTDDS_LATENCY_TAG].begin().value().begin().value()[DATA_VALUE_DATA_TAG]);
+
+    // last_reported_data_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DATA_COUNT_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_DATA_COUNT_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_gap_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_GAP_COUNT_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_GAP_COUNT_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_heartbeat_count
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_HEARTBEAT_COUNT_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_HEARTBEAT_COUNT_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_resent_datas
+    check_is_string_int(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RESENT_DATA_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAWRITER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_RESENT_DATA_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // --------------------- Datareaders ---------------------
+
+    // subscription_throughput
+    check_is_string_int(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_SUBSCRIPTION_THROUGHPUT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_double(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_SUBSCRIPTION_THROUGHPUT_TAG].begin().value()[DATA_VALUE_DATA_TAG]);
+
+    // acknack_count
+    check_is_string_int(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_ACKNACK_COUNT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_ACKNACK_COUNT_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // nackfrag_count
+    check_is_string_int(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NACKFRAG_COUNT_TAG].begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NACKFRAG_COUNT_TAG].begin().value()[DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_acknack_count
+    check_is_string_int(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_ACKNACK_COUNT_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_ACKNACK_COUNT_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // last_reported_nackfrag_count
+    check_is_string_int(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NACKFRAG_COUNT_LAST_REPORTED_TAG][DATA_VALUE_SRC_TIME_TAG]);
+    check_is_uint(dump, dump[DATAREADER_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NACKFRAG_COUNT_LAST_REPORTED_TAG][DATA_VALUE_COUNT_TAG]);
+
+    // --------------------- Locators ---------------------
+
+    // network_latency_per_locator
+    check_is_string_int(dump, dump[LOCATOR_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NETWORK_LATENCY_TAG].begin().value().begin().value()[DATA_VALUE_SRC_TIME_TAG]);
+    check_is_double(dump, dump[LOCATOR_CONTAINER_TAG].begin().value()[DATA_VALUE_DATA_TAG]
+            [DATA_KIND_NETWORK_LATENCY_TAG].begin().value().begin().value()[DATA_VALUE_DATA_TAG]);
+
+}
+
 
 class DataBaseTest : public Database
 {
 public:
-
-    const std::map<EntityId, std::shared_ptr<Host>>& hosts()
-    {
-        return hosts_;
-    }
-
-    const std::map<EntityId, std::shared_ptr<User>>& users()
-    {
-        return users_;
-    }
-
-    const std::map<EntityId, std::shared_ptr<Process>>& processes()
-    {
-        return processes_;
-    }
 
     const std::map<EntityId, std::shared_ptr<Domain>>& domains()
     {
         return domains_;
     }
 
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<Topic>>>& topics()
-    {
-        return topics_;
-    }
-
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<DomainParticipant>>>& participants()
-    {
-        return participants_;
-    }
-
-    const std::map<EntityId, std::shared_ptr<Locator>>& locators()
-    {
-        return locators_;
-    }
-
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<Locator>>>& locators_by_participant()
-    {
-        return locators_by_participant_;
-    }
-
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<DomainParticipant>>>& participants_by_locator()
-    {
-        return participants_by_locator_;
-    }
-
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<Domain>>>& domains_by_process()
-    {
-        return domains_by_process_;
-    }
-
-    const std::map<EntityId, std::map<EntityId, std::shared_ptr<Process>>>& processes_by_domain()
-    {
-        return processes_by_domain_;
-    }
-
-    template<typename T>
-    std::map<EntityId, std::map<EntityId, std::shared_ptr<T>>>& get_dds_endpoints()
-    {
-        return dds_endpoints<T>();
-    }
-
-    Qos test_qos = {
-        {"available_builtin_endpoints", 3135},
-        {"lease_duration", {
-             {"nanoseconds", 0},
-             {"seconds", 3}
-         }},
-        {"properties", {
-             {
-                 {"name", "PARTICIPANT_TYPE"},
-                 {"value", "CLIENT"}
-             },
-             {
-                 {"name", "DS_VERSION"},
-                 {"value", "2.0"}
-             }
-         }},
-        {"user_data", "656e636c6176653d2f3b00"},
-        {"vendor_id", "010f"}
-    };
-
-    const std::atomic<int64_t>& next_id()
-    {
-        return next_id_;
-    }
-
 };
 
-/**
- * @brief Fixture for the get_entities method tests
- *
- * \c get_entities retrieves all the entities of a given kind that are reachable from a given entity.
- * The casuistry for this functionality is rather complex,
- * and the tests need a populated database with several entity combinations in order to be able to
- * test all this casuistry.
- *
- * This fixture populates a database with several entities and several relations among them.
- * Since the tests need to know these relations, each entity is given a unique identifier within the fixture,
- * and the fixture keeps all these entities on a map,
- * for the test to be able to get them and check the result of the execution.
- * This identifier is independent of the EntityId given by the database,
- * and is totally under control of the fixture and the tests.
- * This is important to keep the tests stable even if the Database implementation to assign an EntityId changes.
- *
- * The following Json object describes the populated database objects and their given unique identifiers:
- *
- * \code
- * {
- *  "host1": {
- *     "ID": 1,
- *     "users": []
- *   },
- *   "host2": {
- *     "ID": 2,
- *     "users": ["user1", "user2"]
- *   },
- *   "user1": {
- *     "ID": 3,
- *     "processes": []
- *   },
- *   "user2": {
- *     "ID": 4,
- *     "processes": ["process1", "process2"]
- *   },
- *   "process1": {
- *     "ID": 5,
- *     "participants": []
- *   },
- *   "process2": {
- *     "ID": 6,
- *     "participants": ["participant1", "participant2"]
- *   },
- *   "domain1": {
- *     "ID": 7,
- *     "participants": [],
- *     "topics": []
- *   },
- *   "domain2": {
- *     "ID": 8,
- *     "participants": ["participant1", "participant2"],
- *     "topics": ["topic1", "topic2"]
- *   },
- *   "participant1": {
- *     "ID": 9,
- *     "datareaders": [],
- *     "datawriters": []
- *   },
- *   "participant2": {
- *     "ID": 10,
- *     "datareaders": ["datareader1", "datareader2"],
- *     "datawriters": ["datawriter1", "datawriter2"]
- *   },
- *   "topic1": {
- *     "ID": 11,
- *     "datareaders": [],
- *     "datawriters": []
- *   },
- *   "topic2": {
- *     "ID": 12,
- *     "datareaders": ["datareader1", "datareader2"],
- *     "datawriters": ["datawriter1", "datawriter2"]
- *   },
- *   "datareader1": {
- *     "ID": 13,
- *     "locators": ["reader_locator1"]
- *   }
- *   "datareader2": {
- *     "ID": 14,
- *     "locators": ["reader_locator1", "reader_locator2"]
- *   }
- *   "datawriter1": {
- *     "ID": 15,
- *     "locators": ["writer_locator1"]
- *   }
- *   "datawriter2": {
- *     "ID": 16,
- *     "locators": ["writer_locator1", "writer_locator2"]
- *   }
- *   "reader_locator1": {
- *     "ID": 17
- *   }
- *   "reader_locator2": {
- *     "ID": 18
- *   }
- *   "writer_locator1": {
- *     "ID": 19
- *   }
- *   "writer_locator2": {
- *     "ID": 20
- *   }
- * }
- * \endcode
- *
- *
- * Parameteres to the tests are:
- *  - std::get<0>(GetParam()) The EntityKind we are looking for
- *  - std::get<1>(GetParam()) The unique identifier of the origin Entity, as given by the fixture/testing
- *  - std::get<2>(GetParam()) A list containing the unique identifiers of the entities expected in the result
- */
-class database_load_tests : public ::testing::Test
+// Test the load of a database with wrong references
+TEST(database_load_tests, load_wrong_references)
 {
-protected:
+    // Read JSON
+    DatabaseDump dump = load_file(SIMPLE_DUMP_FILE);
 
-    typedef uint32_t TestId;
-
-    void SetUp()
+    // Host reference user that no exists
     {
-        populate_database();
+        DatabaseDump dumpCopy = dump;
+
+        DatabaseDump container = DatabaseDump::object();
+
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "host_1";
+        entity_info[USER_CONTAINER_TAG] = DatabaseDump::array({"200"});
+        container["100"] = entity_info;
+
+        dumpCopy[HOST_CONTAINER_TAG].insert(container.begin(), container.end());
+
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
     }
 
-    void populate_database()
+    // User reference host that no exists
     {
-        auto host1 = std::make_shared<Host>("host1");
-        db.insert(host1);
-        entities[1] = host1;
+        DatabaseDump dumpCopy = dump;
 
-        auto host2 = std::make_shared<Host>("host2");
-        db.insert(host2);
-        entities[2] = host2;
+        DatabaseDump container = DatabaseDump::object();
 
-        auto user1 = std::make_shared<User>("user1", host2);
-        db.insert(user1);
-        entities[3] = user1;
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "user_1";
+        entity_info[HOST_ENTITY_TAG] = "200";
+        entity_info[PROCESS_CONTAINER_TAG] = DatabaseDump::array();
+        container["100"] = entity_info;
 
-        auto user2 = std::make_shared<User>("user2", host2);
-        db.insert(user2);
-        entities[4] = user2;
+        dumpCopy[USER_CONTAINER_TAG].insert(container.begin(), container.end());
 
-        auto process1 = std::make_shared<Process>("process1", "123", user2);
-        db.insert(process1);
-        entities[5] = process1;
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
+    }
 
-        auto process2 = std::make_shared<Process>("process2", "345", user2);
-        db.insert(process2);
-        entities[6] = process2;
+    // Host reference user, which one is not referencing the host
+    {
+        DatabaseDump dumpCopy = dump;
 
-        auto domain1 = std::make_shared<Domain>("domain1");
-        db.insert(domain1);
-        entities[7] = domain1;
+        DatabaseDump container = DatabaseDump::object();
 
-        auto domain2 = std::make_shared<Domain>("domain2");
-        db.insert(domain2);
-        entities[8] = domain2;
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "host_1";
+        entity_info[USER_CONTAINER_TAG] = dumpCopy[USER_CONTAINER_TAG].begin().key();
+        container["100"] = entity_info;
 
-        auto participant1 = std::make_shared<DomainParticipant>("participant1", "qos", "1.2.3.4", nullptr, domain2);
-        db.insert(participant1);
-        db.link_participant_with_process(participant1->id, process2->id);
-        entities[9] = participant1;
+        dumpCopy[HOST_CONTAINER_TAG].insert(container.begin(), container.end());
 
-        auto participant2 = std::make_shared<DomainParticipant>("participant2", "qos", "5.6.7.8", nullptr, domain2);
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
+    }
+
+    // User reference host, which one is not referencing the user
+    {
+        DatabaseDump dumpCopy = dump;
+
+        DatabaseDump container = DatabaseDump::object();
+
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "user_1";
+        entity_info[HOST_ENTITY_TAG] = dumpCopy[HOST_CONTAINER_TAG].begin().key();
+        entity_info[PROCESS_CONTAINER_TAG] = DatabaseDump::array();
+        container["100"] = entity_info;
+
+        dumpCopy[USER_CONTAINER_TAG].insert(container.begin(), container.end());
+
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
+    }
+
+    // Locator reference datawriter that no exists
+    {
+        DatabaseDump dumpCopy = dump;
+
+        DatabaseDump container = DatabaseDump::object();
+
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "locator_1";
+        entity_info[DATAWRITER_CONTAINER_TAG] = DatabaseDump::array({"200"});
+        entity_info[DATAREADER_CONTAINER_TAG] =
+                dumpCopy[LOCATOR_CONTAINER_TAG][dumpCopy[LOCATOR_CONTAINER_TAG].begin().key()]
+                [DATAREADER_CONTAINER_TAG];
+        entity_info[DATA_CONTAINER_TAG] =
+                dumpCopy[LOCATOR_CONTAINER_TAG][dumpCopy[LOCATOR_CONTAINER_TAG].begin().key()]
+                [DATA_CONTAINER_TAG];
+
+        container["100"] = entity_info;
+
+        dumpCopy[LOCATOR_CONTAINER_TAG].insert(container.begin(), container.end());
+
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
+    }
+
+    // Locator reference datawriter, which one is not referencing the locator
+    {
+        DatabaseDump dumpCopy = dump;
+
+        DatabaseDump container = DatabaseDump::object();
+
+        DatabaseDump entity_info = DatabaseDump::object();
+        entity_info[NAME_INFO_TAG] = "locator_1";
+        entity_info[DATAWRITER_CONTAINER_TAG] = DatabaseDump::array({dumpCopy[DATAWRITER_CONTAINER_TAG].begin().key()});
+        entity_info[DATAREADER_CONTAINER_TAG] =
+                dumpCopy[LOCATOR_CONTAINER_TAG][dumpCopy[LOCATOR_CONTAINER_TAG].begin().key()]
+                [DATAREADER_CONTAINER_TAG];
+        entity_info[DATA_CONTAINER_TAG] =
+                dumpCopy[LOCATOR_CONTAINER_TAG][dumpCopy[LOCATOR_CONTAINER_TAG].begin().key()]
+                [DATA_CONTAINER_TAG];
+
+        container["100"] = entity_info;
+
+        dumpCopy[LOCATOR_CONTAINER_TAG].insert(container.begin(), container.end());
+
+        Database db;
+        ASSERT_THROW(db.load_database(dumpCopy), CorruptedFile);
+    }
+
+    // Participant not referencing any process
+    {
+        DataBaseTest db;
+        db.load_database(dump);
+
+        std::map<EntityId, std::shared_ptr<Domain>> domains = db.domains();
+        std::shared_ptr<Domain> domain = domains[EntityId(std::stoi(dump[DOMAIN_CONTAINER_TAG].begin().key()))];
+        auto participant2 = std::make_shared<DomainParticipant>("participant2", "qos", "5.6.7.8", nullptr, domain);
         db.insert(participant2);
-        db.link_participant_with_process(participant2->id, process2->id);
-        entities[10] = participant2;
 
-        auto topic1 = std::make_shared<Topic>("topic1", "type", domain2);
-        db.insert(topic1);
-        entities[11] = topic1;
-
-        auto topic2 = std::make_shared<Topic>("topic2", "type", domain2);
-        db.insert(topic2);
-        entities[12] = topic2;
-
-        auto datareader1 = std::make_shared<DataReader>("datareader1", "qos", "11.12.13.14", participant2, topic2);
-        auto reader_locator1 = std::make_shared<Locator>("reader_locator1");
-        reader_locator1->id = db.generate_entity_id();
-        datareader1->locators[reader_locator1->id] = reader_locator1;
-        db.insert(datareader1);
-        entities[13] = datareader1;
-        entities[17] = reader_locator1;
-
-        auto datareader2 = std::make_shared<DataReader>("datareader2", "qos", "15.16.17.18", participant2, topic2);
-        auto reader_locator2 = std::make_shared<Locator>("reader_locator2");
-        reader_locator2->id = db.generate_entity_id();
-        datareader2->locators[reader_locator1->id] = reader_locator1;
-        datareader2->locators[reader_locator2->id] = reader_locator2;
-        db.insert(datareader2);
-        entities[14] = datareader2;
-        entities[18] = reader_locator2;
-
-        auto datawriter1 = std::make_shared<DataWriter>("datawriter1", "qos", "21.22.23.24", participant2, topic2);
-        auto writer_locator1 = std::make_shared<Locator>("writer_locator1");
-        writer_locator1->id = db.generate_entity_id();
-        datawriter1->locators[writer_locator1->id] = writer_locator1;
-        db.insert(datawriter1);
-        entities[15] = datawriter1;
-        entities[19] = writer_locator1;
-
-        auto datawriter2 = std::make_shared<DataWriter>("datawriter2", "qos", "25.26.27.28", participant2, topic2);
-        auto writer_locator2 = std::make_shared<Locator>("writer_locator2");
-        writer_locator2->id = db.generate_entity_id();
-        datawriter2->locators[writer_locator1->id] = writer_locator1;
-        datawriter2->locators[writer_locator2->id] = writer_locator2;
-        db.insert(datawriter2);
-        entities[16] = datawriter2;
-        entities[20] = writer_locator2;
-
-        // Insert datas on domain2
-
-        // participants
-        {
-            EntityId domainId = participant2->domain->id;
-            EntityId entityId = participant2->id;
-
-            // discovery_time
-            {
-                DiscoveryTimeSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(2));
-                sample.time = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.remote_entity = EntityId(entityId);
-                sample.discovered = true;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // pdp_packets
-            {
-                PdpCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(0));
-                sample.count = 1;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // edp_packets
-            {
-                EdpCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // rtps_packets_sent
-            {
-                RtpsPacketsSentSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-                sample.remote_locator = EntityId(reader_locator2->id);
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // rtps_bytes_sent
-            {
-                RtpsBytesSentSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-                sample.magnitude_order = 0;
-                sample.remote_locator = EntityId(reader_locator2->id);
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // rtps_packets_lost
-            {
-                RtpsPacketsLostSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-                sample.remote_locator = EntityId(reader_locator2->id);
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // rtps_bytes_lost
-            {
-                RtpsBytesLostSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-                sample.magnitude_order = 0;
-                sample.remote_locator = EntityId(reader_locator2->id);
-
-                db.insert(domainId, entityId, sample);
-            }
-        }
-
-        // datawriters
-        {
-            EntityId domainId = datawriter2->participant->domain->id;
-            EntityId entityId = datawriter2->id;
-
-            // publication_throughput
-            {
-                PublicationThroughputSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.data = 1.1;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // resent_datas
-            {
-                ResentDataSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // heartbeat_count
-            {
-                HeartbeatCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // gap_count
-            {
-                GapCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // data_count
-            {
-                DataCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // samples_datas
-            {
-                SampleDatasCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-                sample.sequence_number = process2->id.value();
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // history2history_latency
-            {
-                HistoryLatencySample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.data = 1.1;
-                sample.reader = (EntityId(reader_locator2->id));
-
-                db.insert(domainId, entityId, sample);
-            }
-        }
-
-        // datareaders
-        {
-            EntityId domainId = datareader2->participant->domain->id;
-            EntityId entityId = datareader2->id;
-
-            // subscription_throughput
-            {
-                SubscriptionThroughputSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.data = 1.1;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // acknack_count
-            {
-                AcknackCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-
-            // nackfrag_count
-            {
-                NackfragCountSample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.count = 0;
-
-                db.insert(domainId, entityId, sample);
-            }
-        }
-
-        // locators
-        {
-            // writer_locator
-            {
-                EntityId entityId = writer_locator2->id;
-
-                NetworkLatencySample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.data = 1.1;
-                sample.remote_locator = EntityId(entityId);
-
-                db.insert(EntityId::invalid(), entityId, sample);
-            }
-
-            // reader_locator
-            {
-                EntityId entityId = reader_locator2->id;
-
-                NetworkLatencySample sample;
-
-                sample.src_ts = std::chrono::system_clock::time_point(std::chrono::steady_clock::duration(1));
-                sample.data = 1.1;
-                sample.remote_locator = EntityId(entityId);
-
-                db.insert(EntityId::invalid(), entityId, sample);
-            }
-
-        }
+        DatabaseDump dumpCopy = db.dump_database();
+        DataBaseTest newDb;
+        ASSERT_NO_THROW(newDb.load_database(dumpCopy); );
     }
-
-    DataBaseTest db;
-    std::map<TestId, std::shared_ptr<const Entity>> entities;
-};
-
-template <typename Map>
-bool key_compare (
-        Map const& lhs,
-        Map const& rhs)
-{
-    return lhs.size() == rhs.size()
-           && std::equal(lhs.begin(), lhs.end(), rhs.begin(),
-                   [](auto a, auto b)
-                   {
-                       return a.first == b.first;
-                   });
 }
 
-template <typename Map>
-bool map_compare (
-        Map const& lhs,
-        Map const& rhs)
-{
-    // No predicate needed because there is operator== for pairs already.
-    return lhs.size() == rhs.size()
-           && std::equal(lhs.begin(), lhs.end(),
-                   rhs.begin());
-}
-
-// Test the load of a database
-TEST_F(database_load_tests, load_database)
-{
-    // Dump of the database with inserted datas
-    DatabaseDump dump = db.dump_database();
-
-    // Create db_loaded
-    DataBaseTest db_loaded;
-    db_loaded.load_database(dump);
-
-    // ------------------ Compare map keys ------------------------
-
-    // Compare ID of all maps of both databases.
-    // If at() throw an exception, the db_loaded does not contain one Entity which should contain
-
-    // hosts
-    ASSERT_TRUE(key_compare(db.hosts(), db_loaded.hosts()));
-
-    // users
-    ASSERT_TRUE(key_compare(db.users(), db_loaded.users()));
-
-    // processes
-    ASSERT_TRUE(key_compare(db.processes(), db_loaded.processes()));
-
-    // domains
-    ASSERT_TRUE(key_compare(db.domains(), db_loaded.domains()));
-
-    // topics
-    ASSERT_TRUE(key_compare(db.topics(), db_loaded.topics()));
-    for (auto it = db.topics().cbegin(); it != db.topics().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.topics().at(it->first)));
-    }
-
-    // participants
-    ASSERT_TRUE(key_compare(db.participants(), db_loaded.participants()));
-    for (auto it = db.participants().cbegin(); it != db.participants().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.participants().at(it->first)));
-    }
-
-    // locators
-    ASSERT_TRUE(key_compare(db.locators(), db_loaded.locators()));
-
-    // DataWriter
-    ASSERT_TRUE(key_compare(db.get_dds_endpoints<DataWriter>(), db_loaded.get_dds_endpoints<DataWriter>()));
-    for (auto it = db.get_dds_endpoints<DataWriter>().cbegin(); it != db.get_dds_endpoints<DataWriter>().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.get_dds_endpoints<DataWriter>().at(it->first)));
-    }
-
-    // DataReader
-    ASSERT_TRUE(key_compare(db.get_dds_endpoints<DataReader>(), db_loaded.get_dds_endpoints<DataReader>()));
-    for (auto it = db.get_dds_endpoints<DataReader>().cbegin(); it != db.get_dds_endpoints<DataReader>().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.get_dds_endpoints<DataReader>().at(it->first)));
-    }
-
-    // locators_by_participant
-    ASSERT_TRUE(key_compare(db.locators_by_participant(), db_loaded.locators_by_participant()));
-    for (auto it = db.locators_by_participant().cbegin(); it != db.locators_by_participant().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.locators_by_participant().at(it->first)));
-    }
-
-    // participants_by_locator
-    ASSERT_TRUE(key_compare(db.participants_by_locator(), db_loaded.participants_by_locator()));
-    for (auto it = db.participants_by_locator().cbegin(); it != db.participants_by_locator().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.participants_by_locator().at(it->first)));
-    }
-
-    // domains_by_process
-    ASSERT_TRUE(key_compare(db.domains_by_process(), db_loaded.domains_by_process()));
-    for (auto it = db.domains_by_process().cbegin(); it != db.domains_by_process().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.domains_by_process().at(it->first)));
-    }
-
-    // processes_by_domain
-    ASSERT_TRUE(key_compare(db.processes_by_domain(), db_loaded.processes_by_domain()));
-    for (auto it = db.processes_by_domain().cbegin(); it != db.processes_by_domain().cend(); ++it)
-    {
-        ASSERT_TRUE(key_compare(it->second, db_loaded.processes_by_domain().at(it->first)));
-    }
-
-    // ------------------ Compare data ------------------------
-
-    // Participants
-    for (auto domainIt = db.participants().cbegin(); domainIt != db.participants().cend(); ++domainIt)
-    {
-        for (auto it = db.participants().at(domainIt->first).cbegin();
-                it != db.participants().at(domainIt->first).cend(); ++it)
-        {
-            DomainParticipantData insertedData =  db.participants().at(domainIt->first).at(it->first)->data;
-            DomainParticipantData loadedData =  db_loaded.participants().at(domainIt->first).at(it->first)->data;
-
-            ASSERT_TRUE(map_compare(insertedData.rtps_packets_sent, loadedData.rtps_packets_sent));
-            ASSERT_TRUE(map_compare(insertedData.last_reported_rtps_packets_sent_count,
-                    loadedData.last_reported_rtps_packets_sent_count));
-            ASSERT_TRUE(map_compare(insertedData.rtps_bytes_sent, loadedData.rtps_bytes_sent));
-            ASSERT_TRUE(map_compare(insertedData.last_reported_rtps_bytes_sent_count,
-                    loadedData.last_reported_rtps_bytes_sent_count));
-            ASSERT_TRUE(map_compare(insertedData.rtps_packets_lost, loadedData.rtps_packets_lost));
-            ASSERT_TRUE(map_compare(insertedData.last_reported_rtps_packets_lost_count,
-                    loadedData.last_reported_rtps_packets_lost_count));
-            ASSERT_TRUE(map_compare(insertedData.rtps_bytes_lost, loadedData.rtps_bytes_lost));
-            ASSERT_TRUE(map_compare(insertedData.last_reported_rtps_bytes_lost_count,
-                    loadedData.last_reported_rtps_bytes_lost_count));
-            ASSERT_TRUE(map_compare(insertedData.discovered_entity, loadedData.discovered_entity));
-            ASSERT_TRUE(insertedData.pdp_packets == loadedData.pdp_packets);
-            ASSERT_TRUE(insertedData.last_reported_pdp_packets == loadedData.last_reported_pdp_packets);
-            ASSERT_TRUE(insertedData.edp_packets == loadedData.edp_packets);
-            ASSERT_TRUE(insertedData.last_reported_edp_packets == loadedData.last_reported_edp_packets);
-        }
-    }
-
-    // DataWriter
-    for (auto domainIt = db.get_dds_endpoints<DataWriter>().cbegin();
-            domainIt != db.get_dds_endpoints<DataWriter>().cend(); ++domainIt)
-    {
-        for (auto it = db.get_dds_endpoints<DataWriter>().at(domainIt->first).cbegin();
-                it != db.get_dds_endpoints<DataWriter>().at(domainIt->first).cend(); ++it)
-        {
-            DataWriterData insertedData =
-                    db.get_dds_endpoints<DataWriter>().at(domainIt->first).at(it->first)->data;
-            DataWriterData loadedData =
-                    db_loaded.get_dds_endpoints<DataWriter>().at(domainIt->first).at(it->first)->data;
-
-            ASSERT_TRUE(insertedData.publication_throughput == loadedData.publication_throughput);
-            ASSERT_TRUE(insertedData.resent_datas == loadedData.resent_datas);
-            ASSERT_TRUE(insertedData.last_reported_resent_datas == loadedData.last_reported_resent_datas);
-            ASSERT_TRUE(insertedData.heartbeat_count == loadedData.heartbeat_count);
-            ASSERT_TRUE(insertedData.last_reported_heartbeat_count == loadedData.last_reported_heartbeat_count);
-            ASSERT_TRUE(insertedData.gap_count == loadedData.gap_count);
-            ASSERT_TRUE(insertedData.last_reported_gap_count == loadedData.last_reported_gap_count);
-            ASSERT_TRUE(insertedData.data_count == loadedData.data_count);
-            ASSERT_TRUE(insertedData.last_reported_data_count == loadedData.last_reported_data_count);
-            ASSERT_TRUE(map_compare(insertedData.sample_datas, loadedData.sample_datas));
-            ASSERT_TRUE(map_compare(insertedData.history2history_latency, loadedData.history2history_latency));
-        }
-    }
-
-    // DataReader
-    for (auto domainIt = db.get_dds_endpoints<DataReader>().cbegin();
-            domainIt != db.get_dds_endpoints<DataReader>().cend(); ++domainIt)
-    {
-        for (auto it = db.get_dds_endpoints<DataReader>().at(domainIt->first).cbegin();
-                it != db.get_dds_endpoints<DataReader>().at(domainIt->first).cend(); ++it)
-        {
-            DataReaderData insertedData =
-                    db.get_dds_endpoints<DataReader>().at(domainIt->first).at(it->first)->data;
-            DataReaderData loadedData =
-                    db_loaded.get_dds_endpoints<DataReader>().at(domainIt->first).at(it->first)->data;
-
-            ASSERT_TRUE(insertedData.subscription_throughput == loadedData.subscription_throughput);
-            ASSERT_TRUE(insertedData.acknack_count == loadedData.acknack_count);
-            ASSERT_TRUE(insertedData.last_reported_acknack_count == loadedData.last_reported_acknack_count);
-            ASSERT_TRUE(insertedData.nackfrag_count == loadedData.nackfrag_count);
-            ASSERT_TRUE(insertedData.last_reported_nackfrag_count == loadedData.last_reported_nackfrag_count);
-
-        }
-    }
-
-    // Locator
-    for (auto it = db.locators().cbegin(); it != db.locators().cend(); ++it)
-    {
-        LocatorData insertedData = db.locators().at(it->first)->data;
-        LocatorData loadedData = db_loaded.locators().at(it->first)->data;
-
-        ASSERT_TRUE(map_compare(insertedData.network_latency_per_locator, loadedData.network_latency_per_locator));
-    }
-
-    // ------------------ Compare entities ------------------------
-
-
-    // Host
-    for (auto insertedIt = db.hosts().cbegin(), loadedIt = db_loaded.hosts().cbegin();
-            insertedIt != db.hosts().cend() && loadedIt != db_loaded.hosts().cend(); insertedIt++, loadedIt++)
-    {
-        std::shared_ptr<Host> insertedEntity = insertedIt->second;
-        std::shared_ptr<Host> loadedEntity = loadedIt->second;
-
-        ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                insertedEntity->name == loadedEntity->name);
-        ASSERT_TRUE(key_compare(insertedEntity->users, loadedEntity->users));
-    }
-
-    // Users
-    for (auto insertedIt = db.users().cbegin(), loadedIt = db_loaded.users().cbegin();
-            insertedIt != db.users().cend() && loadedIt != db_loaded.users().cend(); insertedIt++, loadedIt++)
-    {
-        std::shared_ptr<User> insertedEntity = insertedIt->second;
-        std::shared_ptr<User> loadedEntity = loadedIt->second;
-
-        ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                insertedEntity->name == loadedEntity->name);
-        ASSERT_TRUE(insertedEntity->host->id == loadedEntity->host->id);
-        ASSERT_TRUE(key_compare(insertedEntity->processes, loadedEntity->processes));
-    }
-
-    // Processes
-    for (auto insertedIt = db.processes().cbegin(), loadedIt = db_loaded.processes().cbegin();
-            insertedIt != db.processes().cend() && loadedIt != db_loaded.processes().cend(); insertedIt++, loadedIt++)
-    {
-        std::shared_ptr<Process> insertedEntity = insertedIt->second;
-        std::shared_ptr<Process> loadedEntity = loadedIt->second;
-
-        ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                insertedEntity->name == loadedEntity->name);
-
-        ASSERT_TRUE(insertedEntity->pid == loadedEntity->pid);
-        ASSERT_TRUE(insertedEntity->user->id == loadedEntity->user->id);
-        ASSERT_TRUE(key_compare(insertedEntity->participants, loadedEntity->participants));
-    }
-
-    // Locators
-    for (auto insertedIt = db.locators().cbegin(), loadedIt = db_loaded.locators().cbegin();
-            insertedIt != db.locators().cend() && loadedIt != db_loaded.locators().cend(); insertedIt++, loadedIt++)
-    {
-        std::shared_ptr<Locator> insertedEntity = insertedIt->second;
-        std::shared_ptr<Locator> loadedEntity = loadedIt->second;
-
-        ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                insertedEntity->name == loadedEntity->name);
-
-        ASSERT_TRUE(key_compare(insertedEntity->data_readers, loadedEntity->data_readers));
-        ASSERT_TRUE(key_compare(insertedEntity->data_writers, loadedEntity->data_writers));
-    }
-
-    // Domains
-    for (auto insertedIt = db.domains().cbegin(), loadedIt = db_loaded.domains().cbegin();
-            insertedIt != db.domains().cend() && loadedIt != db_loaded.domains().cend(); insertedIt++, loadedIt++)
-    {
-        std::shared_ptr<Domain> insertedEntity = insertedIt->second;
-        std::shared_ptr<Domain> loadedEntity = loadedIt->second;
-
-        ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                insertedEntity->name == loadedEntity->name);
-
-        ASSERT_TRUE(key_compare(insertedEntity->topics, loadedEntity->topics));
-        ASSERT_TRUE(key_compare(insertedEntity->participants, loadedEntity->participants));
-    }
-
-    // Participants
-    for (auto domainIt = db.participants().cbegin(); domainIt != db.participants().cend(); ++domainIt)
-    {
-        for (auto insertedIt = db.participants().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.participants().at(domainIt->first).cbegin();
-                insertedIt != db.participants().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.participants().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<DomainParticipant> insertedEntity = insertedIt->second;
-            std::shared_ptr<DomainParticipant> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->qos == loadedEntity->qos && insertedEntity->guid == loadedEntity->guid);
-
-            ASSERT_TRUE(insertedEntity->process->id == loadedEntity->process->id);
-            ASSERT_TRUE(insertedEntity->domain->id == loadedEntity->domain->id);
-
-            ASSERT_TRUE(key_compare(insertedEntity->data_readers, loadedEntity->data_readers));
-            ASSERT_TRUE(key_compare(insertedEntity->data_writers, loadedEntity->data_writers));
-        }
-    }
-
-    // Datawriters
-    for (auto domainIt = db.get_dds_endpoints<DataWriter>().cbegin(); domainIt !=
-            db.get_dds_endpoints<DataWriter>().cend();
-            ++domainIt)
-    {
-        for (auto insertedIt = db.get_dds_endpoints<DataWriter>().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.get_dds_endpoints<DataWriter>().at(domainIt->first).cbegin();
-                insertedIt != db.get_dds_endpoints<DataWriter>().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.get_dds_endpoints<DataWriter>().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<DataWriter> insertedEntity = insertedIt->second;
-            std::shared_ptr<DataWriter> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->qos == loadedEntity->qos && insertedEntity->guid == loadedEntity->guid);
-
-            ASSERT_TRUE(insertedEntity->participant->id == loadedEntity->participant->id);
-            ASSERT_TRUE(insertedEntity->topic->id == loadedEntity->topic->id);
-
-            ASSERT_TRUE(key_compare(insertedEntity->locators, loadedEntity->locators));
-        }
-    }
-
-    // Datareaders
-    for (auto domainIt = db.get_dds_endpoints<DataReader>().cbegin(); domainIt !=
-            db.get_dds_endpoints<DataReader>().cend();
-            ++domainIt)
-    {
-        for (auto insertedIt = db.get_dds_endpoints<DataReader>().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.get_dds_endpoints<DataReader>().at(domainIt->first).cbegin();
-                insertedIt != db.get_dds_endpoints<DataReader>().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.get_dds_endpoints<DataReader>().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<DataReader> insertedEntity = insertedIt->second;
-            std::shared_ptr<DataReader> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->qos == loadedEntity->qos && insertedEntity->guid == loadedEntity->guid);
-
-            ASSERT_TRUE(insertedEntity->participant->id == loadedEntity->participant->id);
-            ASSERT_TRUE(insertedEntity->topic->id == loadedEntity->topic->id);
-
-            ASSERT_TRUE(key_compare(insertedEntity->locators, loadedEntity->locators));
-        }
-    }
-
-    // Topics
-    for (auto domainIt = db.topics().cbegin(); domainIt != db.topics().cend(); ++domainIt)
-    {
-        for (auto insertedIt = db.topics().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.topics().at(domainIt->first).cbegin();
-                insertedIt != db.topics().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.topics().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<Topic> insertedEntity = insertedIt->second;
-            std::shared_ptr<Topic> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->data_type == loadedEntity->data_type);
-            ASSERT_TRUE(insertedEntity->domain->id == loadedEntity->domain->id);
-
-            ASSERT_TRUE(key_compare(insertedEntity->data_readers, loadedEntity->data_readers));
-            ASSERT_TRUE(key_compare(insertedEntity->data_writers, loadedEntity->data_writers));
-        }
-    }
-
-    // domains_by_process
-    for (auto domainIt = db.domains_by_process().cbegin(); domainIt != db.domains_by_process().cend(); ++domainIt)
-    {
-        for (auto insertedIt = db.domains_by_process().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.domains_by_process().at(domainIt->first).cbegin();
-                insertedIt != db.domains_by_process().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.domains_by_process().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<Domain> insertedEntity = insertedIt->second;
-            std::shared_ptr<Domain> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(key_compare(insertedEntity->topics, loadedEntity->topics));
-            ASSERT_TRUE(key_compare(insertedEntity->participants, loadedEntity->participants));
-        }
-    }
-
-    // processes_by_domain_
-    for (auto domainIt = db.processes_by_domain().cbegin(); domainIt != db.processes_by_domain().cend(); ++domainIt)
-    {
-        for (auto insertedIt = db.processes_by_domain().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.processes_by_domain().at(domainIt->first).cbegin();
-                insertedIt != db.processes_by_domain().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.processes_by_domain().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<Process> insertedEntity = insertedIt->second;
-            std::shared_ptr<Process> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->pid == loadedEntity->pid);
-            ASSERT_TRUE(insertedEntity->user->id == loadedEntity->user->id);
-            ASSERT_TRUE(key_compare(insertedEntity->participants, loadedEntity->participants));
-        }
-    }
-
-    // participants_by_locator
-    for (auto domainIt = db.participants_by_locator().cbegin(); domainIt != db.participants_by_locator().cend();
-            ++domainIt)
-    {
-        for (auto insertedIt = db.participants_by_locator().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.participants_by_locator().at(domainIt->first).cbegin();
-                insertedIt != db.participants_by_locator().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.participants_by_locator().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<DomainParticipant> insertedEntity = insertedIt->second;
-            std::shared_ptr<DomainParticipant> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(insertedEntity->qos == loadedEntity->qos && insertedEntity->guid == loadedEntity->guid);
-
-            ASSERT_TRUE(insertedEntity->process->id == loadedEntity->process->id);
-            ASSERT_TRUE(insertedEntity->domain->id == loadedEntity->domain->id);
-
-            ASSERT_TRUE(key_compare(insertedEntity->data_readers, loadedEntity->data_readers));
-            ASSERT_TRUE(key_compare(insertedEntity->data_writers, loadedEntity->data_writers));
-        }
-    }
-
-    // locators_by_participant
-    for (auto domainIt = db.locators_by_participant().cbegin(); domainIt != db.locators_by_participant().cend();
-            ++domainIt)
-    {
-        for (auto insertedIt = db.locators_by_participant().at(domainIt->first).cbegin(),
-                loadedIt = db_loaded.locators_by_participant().at(domainIt->first).cbegin();
-                insertedIt != db.locators_by_participant().at(domainIt->first).cend() &&
-                loadedIt != db_loaded.locators_by_participant().at(domainIt->first).cend();
-                insertedIt++, loadedIt++)
-        {
-            std::shared_ptr<Locator> insertedEntity = insertedIt->second;
-            std::shared_ptr<Locator> loadedEntity = loadedIt->second;
-
-            ASSERT_TRUE(insertedEntity->id == loadedEntity->id && insertedEntity->kind == loadedEntity->kind &&
-                    insertedEntity->name == loadedEntity->name);
-
-            ASSERT_TRUE(key_compare(insertedEntity->data_readers, loadedEntity->data_readers));
-            ASSERT_TRUE(key_compare(insertedEntity->data_writers, loadedEntity->data_writers));
-        }
-    }
-
-    // Compare next_id_ of both databases
-    ASSERT_EQ(db.next_id(), db_loaded.next_id());
-
-    // Compare dump of both databases
-    ASSERT_EQ(dump, db_loaded.dump_database());
-}
 
 int main(
         int argc,
